\name{chess.rad}
\alias{chess.rad}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Chess (King-Rook vs. King-Pawn) Data Set 
}
\description{
Chess (King-Rook vs. King-Pawn) Data Set from UCI repository
}


\references{
  \url{http://archive.ics.uci.edu/ml/datasets/Chess+(King-Rook+vs.+King-Pawn)}
}
